MonoliticVersion/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
eventlet==0.33.3
psycopg2-binary==2.9.7